Comments
Reworked App.js to route to homepage instead of directing to login.

Changed ProtectedRoute.js to navigate to the homepage instead of login.

Homepage.js added to coding/src/pages.

index.js
	removed strictmode to prevent
build errors.

Added to Components folder
	Navbar folder
		menuitem.js
		navbar.css
		navbar.js
	button.css
	button.js

Edited: App.css to match website fonts and format.

Navbar was added to Homepage/Register/Login.
Note: This is just the start. More can be added. If you look through the code, a logo is technically in from fontawesome. I have an example template of a hero and logo as well. I will continue working on adding a logo next to code-breaker and hero page for the homepage once  the login/register page is reworked (hopefully done with that Wed.). I didn't add the Navbar to teacher/student dashboard or about page, but it can be easily done.

Special Note: Navbar is currently visual only. Clicking on anything in it doesn't #link to anything yet. If anyone can do this easily, feel free.